#2382 separating model classes
