+++ docs/source/models/equations.rst
@@ -1,0 +1,2 @@
+Equations
+---------
