+++ examples/scripts/SPMe.py
@@ -16,7 +16,7 @@ geometry = model.default_geometry
 
 # load parameter values and process model and geometry
 param = model.default_parameter_values
-param.process_model(model)
+model = param.process_model(model)
 param.process_geometry(geometry)
 
 # set mesh
